tests/requirements.txt
selenium>=4.15.2
python-dotenv==1.0.0
pytest==7.4.3
